- works properly now
